♻️ refactor: 폴더 구조 변경 및 피드백 반영
