+++ app/(public)/layout.tsx
@@ -1,5 +1,5 @@
-import BottomNav from "@/src/shared/ui/bottom-nav";
-import Header from "@/src/shared/ui/header";
+import BottomNav from "@/src/shared/layout/bottom-nav";
+import Header from "@/src/shared/layout/header";
 import { ReactNode } from "react";
 
 export default function Layout({ children }: { children: ReactNode }) {
