+++ src/stories/bottom-nav.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import BottomNav from "@/src/shared/ui/bottom-nav";
+import BottomNav from "@/src/shared/layout/bottom-nav";
 
 const meta: Meta<typeof BottomNav> = {
   title: "ui-kit/BottomNav",
